chore: refactor out useDefaultAddress hook
