+++ ui/components/multichain-accounts/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useDefaultAddress } from './useDefaultAddress';
