+++ ui/components/multichain-accounts/multichain-account-network-group-with-copy-icon/multichain-account-network-group-with-copy-icon.tsx
@@ -1,5 +1,4 @@
-import React, { useCallback } from 'react';
-import { useSelector } from 'react-redux';
+import React from 'react';
 import classnames from 'clsx';
 import { type AccountGroupId } from '@metamask/account-api';
 import {
@@ -20,18 +19,9 @@ import { useI18nContext } from '../../../hooks/useI18nContext';
 import { shortenAddress } from '../../../helpers/utils/util';
 // eslint-disable-next-line import-x/no-restricted-paths
 import { normalizeSafeAddress } from '../../../../app/scripts/lib/multichain/address';
-import { getDefaultScopeAndAddressByAccountGroupId } from '../../../selectors/multichain-accounts/account-tree';
-import {
-  getIsDefaultAddressEnabled,
-  getShowDefaultAddressPreference,
-  getDefaultAddressScope,
-} from '../../../selectors/selectors';
 import { MultichainAccountNetworkGroup } from '../multichain-account-network-group';
-import { useCopyToClipboard } from '../../../hooks/useCopyToClipboard';
-import {
-  DEFAULT_ADDRESS_DISPLAY_KEY_BY_SCOPE,
-  DefaultAddressScope,
-} from '../../../../shared/constants/default-address';
+import { DEFAULT_ADDRESS_DISPLAY_KEY_BY_SCOPE } from '../../../../shared/constants/default-address';
+import { useDefaultAddress } from '../hooks/useDefaultAddress';
 
 const MAX_NETWORK_AVATARS = 4;
 
@@ -40,9 +30,9 @@ export type MultichainAccountNetworkGroupWithCopyIconProps = {
 };
 
 /**
- * Displays network avatars with a copy icon. When the showDefaultAddress
- * preference is enabled, also displays the shortened default address.
- * Click copies the default address and shows "Copied" briefly.
+ * Displays network avatars with a copy icon. When a default address
+ * is available, also displays the shortened default address.
+ * Click copies the default address and shows a message.
  *
  * @param options0
  * @param options0.groupId
@@ -51,27 +41,13 @@ export const MultichainAccountNetworkGroupWithCopyIcon = ({
   groupId,
 }: MultichainAccountNetworkGroupWithCopyIconProps) => {
   const t = useI18nContext();
-  const isDefaultAddressEnabled = useSelector(getIsDefaultAddressEnabled);
-  const showDefaultAddressPreference = useSelector(
-    getShowDefaultAddressPreference,
-  );
-  const defaultAddressScope = useSelector(
-    getDefaultAddressScope,
-  ) as DefaultAddressScope;
-  const { defaultAddress } = useSelector((state) =>
-    getDefaultScopeAndAddressByAccountGroupId(state, groupId),
-  );
-  const displayDefaultAddress =
-    isDefaultAddressEnabled && showDefaultAddressPreference && defaultAddress;
-  const [addressCopied, handleCopy] = useCopyToClipboard({
-    clearDelayMs: null,
-  });
-
-  const handleDefaultAddressClick = useCallback(() => {
-    if (defaultAddress) {
-      handleCopy(normalizeSafeAddress(defaultAddress));
-    }
-  }, [defaultAddress, handleCopy]);
+  const {
+    defaultAddress,
+    defaultAddressScope,
+    displayDefaultAddress,
+    addressCopied,
+    handleDefaultAddressClick,
+  } = useDefaultAddress(groupId);
 
   return (
     <Box
@@ -108,7 +84,7 @@ export const MultichainAccountNetworkGroupWithCopyIcon = ({
         >
           {addressCopied
             ? `${t(DEFAULT_ADDRESS_DISPLAY_KEY_BY_SCOPE[defaultAddressScope])} ${t('addressCopied').toLowerCase()}`
-            : shortenAddress(normalizeSafeAddress(defaultAddress))}
+            : shortenAddress(normalizeSafeAddress(defaultAddress as string))}
         </Text>
       )}
       <Icon
